+++ piecrust/tests/callcenter.rs
@@ -8,6 +8,7 @@ use piecrust::{contract_bytecode, ContractData, Error, SessionData, VM};
 use piecrust_uplink::{ContractError, ContractId, RawCall, RawResult};
 
 const OWNER: [u8; 32] = [0u8; 32];
+const LIMIT: u64 = 1_000_000;
 
 #[test]
 pub fn cc_read_counter() -> Result<(), Error> {
@@ -15,21 +16,27 @@ pub fn cc_read_counter() -> Result<(), Error> {
 
     let mut session = vm.session(SessionData::builder())?;
 
-    let counter_id = session
-        .deploy(contract_bytecode!("counter"), ContractData::builder(OWNER))?;
+    let counter_id = session.deploy(
+        contract_bytecode!("counter"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
 
     // read direct
 
-    let value: i64 = session.call(counter_id, "read_value", &())?;
+    let value: i64 = session.call(counter_id, "read_value", &(), LIMIT)?.data;
     assert_eq!(value, 0xfc);
 
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
     // read value through callcenter
-    let value: i64 = session.call(center_id, "query_counter", &counter_id)?;
+    let value: i64 = session
+        .call(center_id, "query_counter", &counter_id, LIMIT)?
+        .data;
     assert_eq!(value, 0xfc);
 
     Ok(())
@@ -41,31 +48,44 @@ pub fn cc_direct() -> Result<(), Error> {
 
     let mut session = vm.session(SessionData::builder())?;
 
-    let counter_id = session
-        .deploy(contract_bytecode!("counter"), ContractData::builder(OWNER))?;
+    let counter_id = session.deploy(
+        contract_bytecode!("counter"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
 
     // read value directly
-    let value: i64 = session.call(counter_id, "read_value", &())?;
+    let value: i64 = session.call(counter_id, "read_value", &(), LIMIT)?.data;
     assert_eq!(value, 0xfc);
 
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
     // read value through callcenter
-    let value: i64 = session.call(center_id, "query_counter", &counter_id)?;
+    let value: i64 = session
+        .call(center_id, "query_counter", &counter_id, LIMIT)?
+        .data;
     assert_eq!(value, 0xfc);
 
     // increment through call center
-    session.call(center_id, "increment_counter", &counter_id)?;
+    session.call::<_, ()>(
+        center_id,
+        "increment_counter",
+        &counter_id,
+        LIMIT,
+    )?;
 
     // read value directly
-    let value: i64 = session.call(counter_id, "read_value", &())?;
+    let value: i64 = session.call(counter_id, "read_value", &(), LIMIT)?.data;
     assert_eq!(value, 0xfd);
 
     // read value through callcenter
-    let value: i64 = session.call(center_id, "query_counter", &counter_id)?;
+    let value: i64 = session
+        .call(center_id, "query_counter", &counter_id, LIMIT)?
+        .data;
     assert_eq!(value, 0xfd);
 
     Ok(())
@@ -80,11 +100,14 @@ pub fn cc_passthrough() -> Result<(), Error> {
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
     let rq = RawCall::new("read_value", ());
 
-    let res: RawCall = session.call(center_id, "query_passthrough", &rq)?;
+    let res: RawCall = session
+        .call(center_id, "query_passthrough", &rq, LIMIT)?
+        .data;
 
     assert_eq!(rq, res);
 
@@ -97,21 +120,28 @@ pub fn cc_delegated_read() -> Result<(), Error> {
 
     let mut session = vm.session(SessionData::builder())?;
 
-    let counter_id = session
-        .deploy(contract_bytecode!("counter"), ContractData::builder(OWNER))?;
+    let counter_id = session.deploy(
+        contract_bytecode!("counter"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
     let rq = RawCall::new("read_value", ());
 
     // read value through callcenter
-    let res = session.call::<_, RawResult>(
-        center_id,
-        "delegate_query",
-        &(counter_id, rq),
-    )?;
+    let res = session
+        .call::<_, RawResult>(
+            center_id,
+            "delegate_query",
+            &(counter_id, rq),
+            LIMIT,
+        )?
+        .data;
 
     let value: i64 = res.cast();
 
@@ -129,17 +159,26 @@ pub fn cc_delegated_write() -> Result<(), Error> {
     let rt = RawCall::new("increment", ());
 
     let mut session = vm.session(SessionData::builder())?;
-    let counter_id = session
-        .deploy(contract_bytecode!("counter"), ContractData::builder(OWNER))?;
+    let counter_id = session.deploy(
+        contract_bytecode!("counter"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
-    session.call(center_id, "delegate_transaction", &(counter_id, rt))?;
+    session.call::<_, ()>(
+        center_id,
+        "delegate_transaction",
+        &(counter_id, rt),
+        LIMIT,
+    )?;
 
     // read value directly
-    let value: i64 = session.call(counter_id, "read_value", &())?;
+    let value: i64 = session.call(counter_id, "read_value", &(), LIMIT)?.data;
     assert_eq!(value, 0xfd);
 
     Ok(())
@@ -154,11 +193,13 @@ pub fn cc_self() -> Result<(), Error> {
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
     // am i calling myself
-    let calling_self: bool =
-        session.call(center_id, "calling_self", &center_id)?;
+    let calling_self: bool = session
+        .call(center_id, "calling_self", &center_id, LIMIT)?
+        .data;
     assert!(calling_self);
 
     Ok(())
@@ -173,10 +214,11 @@ pub fn cc_caller() -> Result<(), Error> {
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
     let value: Result<bool, ContractError> =
-        session.call(center_id, "call_self", &())?;
+        session.call(center_id, "call_self", &(), LIMIT)?.data;
 
     assert!(value.expect("should succeed"));
 
@@ -192,9 +234,11 @@ pub fn cc_caller_uninit() -> Result<(), Error> {
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
-    let caller: ContractId = session.call(center_id, "return_caller", &())?;
+    let caller: ContractId =
+        session.call(center_id, "return_caller", &(), LIMIT)?.data;
     assert_eq!(caller, ContractId::uninitialized());
 
     Ok(())
@@ -209,9 +253,11 @@ pub fn cc_self_id() -> Result<(), Error> {
     let center_id = session.deploy(
         contract_bytecode!("callcenter"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
-    let value: ContractId = session.call(center_id, "return_self_id", &())?;
+    let value: ContractId =
+        session.call(center_id, "return_self_id", &(), LIMIT)?.data;
     assert_eq!(value, center_id);
 
     Ok(())