+++ piecrust/tests/fibonacci.rs
@@ -7,6 +7,7 @@
 use piecrust::{contract_bytecode, ContractData, Error, SessionData, VM};
 
 const OWNER: [u8; 32] = [0u8; 32];
+const LIMIT: u64 = 1_000_000;
 
 #[test]
 pub fn fibo() -> Result<(), Error> {
@@ -17,13 +18,14 @@ pub fn fibo() -> Result<(), Error> {
     let id = session.deploy(
         contract_bytecode!("fibonacci"),
         ContractData::builder(OWNER),
+        LIMIT,
     )?;
 
-    assert_eq!(session.call::<u32, u64>(id, "nth", &0)?, 1);
-    assert_eq!(session.call::<u32, u64>(id, "nth", &1)?, 1);
-    assert_eq!(session.call::<u32, u64>(id, "nth", &2)?, 2);
-    assert_eq!(session.call::<u32, u64>(id, "nth", &3)?, 3);
-    assert_eq!(session.call::<u32, u64>(id, "nth", &4)?, 5);
+    assert_eq!(session.call::<u32, u64>(id, "nth", &0, LIMIT)?.data, 1);
+    assert_eq!(session.call::<u32, u64>(id, "nth", &1, LIMIT)?.data, 1);
+    assert_eq!(session.call::<u32, u64>(id, "nth", &2, LIMIT)?.data, 2);
+    assert_eq!(session.call::<u32, u64>(id, "nth", &3, LIMIT)?.data, 3);
+    assert_eq!(session.call::<u32, u64>(id, "nth", &4, LIMIT)?.data, 5);
 
     Ok(())
 }