+++ piecrust/tests/persistence.rs
@@ -7,6 +7,7 @@
 use piecrust::{contract_bytecode, ContractData, Error, SessionData, VM};
 
 const OWNER: [u8; 32] = [0u8; 32];
+const LIMIT: u64 = 1_000_000;
 
 #[test]
 fn session_commits_persistence() -> Result<(), Error> {
@@ -21,15 +22,24 @@ fn session_commits_persistence() -> Result<(), Error> {
         id_1 = session.deploy(
             contract_bytecode!("counter"),
             ContractData::builder(OWNER),
+            LIMIT,
         )?;
-        id_2 = session
-            .deploy(contract_bytecode!("box"), ContractData::builder(OWNER))?;
+        id_2 = session.deploy(
+            contract_bytecode!("box"),
+            ContractData::builder(OWNER),
+            LIMIT,
+        )?;
 
-        session.call::<(), ()>(id_1, "increment", &())?;
-        session.call::<i16, ()>(id_2, "set", &0x11)?;
-        assert_eq!(session.call::<(), i64>(id_1, "read_value", &())?, 0xfd);
+        session.call::<_, ()>(id_1, "increment", &(), LIMIT)?;
+        session.call::<i16, ()>(id_2, "set", &0x11, LIMIT)?;
         assert_eq!(
-            session.call::<_, Option<i16>>(id_2, "get", &())?,
+            session.call::<_, i64>(id_1, "read_value", &(), LIMIT)?.data,
+            0xfd
+        );
+        assert_eq!(
+            session
+                .call::<_, Option<i16>>(id_2, "get", &(), LIMIT)?
+                .data,
             Some(0x11)
         );
         commit_1 = session.commit()?;
@@ -39,11 +49,16 @@ fn session_commits_persistence() -> Result<(), Error> {
     {
         let mut session = vm.session(SessionData::builder().base(commit_1))?;
 
-        session.call::<(), ()>(id_1, "increment", &())?;
-        session.call::<i16, ()>(id_2, "set", &0x12)?;
-        assert_eq!(session.call::<(), i64>(id_1, "read_value", &())?, 0xfe);
+        session.call::<_, ()>(id_1, "increment", &(), LIMIT)?;
+        session.call::<i16, ()>(id_2, "set", &0x12, LIMIT)?;
         assert_eq!(
-            session.call::<_, Option<i16>>(id_2, "get", &())?,
+            session.call::<_, i64>(id_1, "read_value", &(), LIMIT)?.data,
+            0xfe
+        );
+        assert_eq!(
+            session
+                .call::<_, Option<i16>>(id_2, "get", &(), LIMIT)?
+                .data,
             Some(0x12)
         );
         commit_2 = session.commit()?;
@@ -54,9 +69,14 @@ fn session_commits_persistence() -> Result<(), Error> {
         let mut session = vm2.session(SessionData::builder().base(commit_1))?;
 
         // check if both contracts' state was restored
-        assert_eq!(session.call::<(), i64>(id_1, "read_value", &())?, 0xfd);
         assert_eq!(
-            session.call::<_, Option<i16>>(id_2, "get", &())?,
+            session.call::<_, i64>(id_1, "read_value", &(), LIMIT)?.data,
+            0xfd
+        );
+        assert_eq!(
+            session
+                .call::<_, Option<i16>>(id_2, "get", &(), LIMIT)?
+                .data,
             Some(0x11)
         );
     }
@@ -66,9 +86,14 @@ fn session_commits_persistence() -> Result<(), Error> {
         let mut session = vm3.session(SessionData::builder().base(commit_2))?;
 
         // check if both contracts' state was restored
-        assert_eq!(session.call::<(), i64>(id_1, "read_value", &())?, 0xfe);
         assert_eq!(
-            session.call::<_, Option<i16>>(id_2, "get", &())?,
+            session.call::<_, i64>(id_1, "read_value", &(), LIMIT)?.data,
+            0xfe
+        );
+        assert_eq!(
+            session
+                .call::<_, Option<i16>>(id_2, "get", &(), LIMIT)?
+                .data,
             Some(0x12)
         );
     }
@@ -79,16 +104,27 @@ fn session_commits_persistence() -> Result<(), Error> {
 fn contracts_persistence() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
     let mut session = vm.session(SessionData::builder())?;
-    let id_1 = session
-        .deploy(contract_bytecode!("counter"), ContractData::builder(OWNER))?;
-    let id_2 = session
-        .deploy(contract_bytecode!("box"), ContractData::builder(OWNER))?;
+    let id_1 = session.deploy(
+        contract_bytecode!("counter"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
+    let id_2 = session.deploy(
+        contract_bytecode!("box"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
 
-    session.call::<(), ()>(id_1, "increment", &())?;
-    session.call::<i16, ()>(id_2, "set", &0x11)?;
-    assert_eq!(session.call::<(), i64>(id_1, "read_value", &())?, 0xfd);
+    session.call::<_, ()>(id_1, "increment", &(), LIMIT)?;
+    session.call::<i16, ()>(id_2, "set", &0x11, LIMIT)?;
     assert_eq!(
-        session.call::<_, Option<i16>>(id_2, "get", &())?,
+        session.call::<_, i64>(id_1, "read_value", &(), LIMIT)?.data,
+        0xfd
+    );
+    assert_eq!(
+        session
+            .call::<_, Option<i16>>(id_2, "get", &(), LIMIT)?
+            .data,
         Some(0x11)
     );
 
@@ -98,9 +134,16 @@ fn contracts_persistence() -> Result<(), Error> {
     let mut session2 = vm2.session(SessionData::builder().base(commit_1))?;
 
     // check if both contracts' state was restored
-    assert_eq!(session2.call::<(), i64>(id_1, "read_value", &())?, 0xfd);
     assert_eq!(
-        session2.call::<_, Option<i16>>(id_2, "get", &())?,
+        session2
+            .call::<_, i64>(id_1, "read_value", &(), LIMIT)?
+            .data,
+        0xfd
+    );
+    assert_eq!(
+        session2
+            .call::<_, Option<i16>>(id_2, "get", &(), LIMIT)?
+            .data,
         Some(0x11)
     );
     Ok(())