+++ piecrust/tests/stack.rs
@@ -7,6 +7,7 @@
 use piecrust::{contract_bytecode, ContractData, Error, SessionData, VM};
 
 const OWNER: [u8; 32] = [0u8; 32];
+const LIMIT: u64 = 1_000_000;
 
 #[test]
 pub fn push_pop() -> Result<(), Error> {
@@ -14,18 +15,21 @@ pub fn push_pop() -> Result<(), Error> {
 
     let mut session = vm.session(SessionData::builder())?;
 
-    let id = session
-        .deploy(contract_bytecode!("stack"), ContractData::builder(OWNER))?;
+    let id = session.deploy(
+        contract_bytecode!("stack"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
 
     let val = 42;
 
-    session.call(id, "push", &val)?;
+    session.call::<_, ()>(id, "push", &val, LIMIT)?;
 
-    let len: u32 = session.call(id, "len", &())?;
+    let len: u32 = session.call(id, "len", &(), LIMIT)?.data;
     assert_eq!(len, 1);
 
-    let popped: Option<i32> = session.call(id, "pop", &())?;
-    let len: i32 = session.call(id, "len", &())?;
+    let popped: Option<i32> = session.call(id, "pop", &(), LIMIT)?.data;
+    let len: i32 = session.call(id, "len", &(), LIMIT)?.data;
 
     assert_eq!(len, 0);
     assert_eq!(popped, Some(val));
@@ -39,27 +43,30 @@ pub fn multi_push_pop() -> Result<(), Error> {
 
     let mut session = vm.session(SessionData::builder())?;
 
-    let id = session
-        .deploy(contract_bytecode!("stack"), ContractData::builder(OWNER))?;
+    let id = session.deploy(
+        contract_bytecode!("stack"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
 
     const N: i32 = 16;
 
     for i in 0..N {
-        session.call(id, "push", &i)?;
-        let len: i32 = session.call(id, "len", &())?;
+        session.call::<_, ()>(id, "push", &i, LIMIT)?;
+        let len: i32 = session.call(id, "len", &(), LIMIT)?.data;
 
         assert_eq!(len, i + 1);
     }
 
     for i in (0..N).rev() {
-        let popped: Option<i32> = session.call(id, "pop", &())?;
-        let len: i32 = session.call(id, "len", &())?;
+        let popped: Option<i32> = session.call(id, "pop", &(), LIMIT)?.data;
+        let len: i32 = session.call(id, "len", &(), LIMIT)?.data;
 
         assert_eq!(len, i);
         assert_eq!(popped, Some(i));
     }
 
-    let popped: Option<i32> = session.call(id, "pop", &())?;
+    let popped: Option<i32> = session.call(id, "pop", &(), LIMIT)?.data;
     assert_eq!(popped, None);
 
     Ok(())