+++ piecrust/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `CallReceipt` struct
+
+### Changed
+
+- Change signature of `Session::deploy` to take `points_limit`
+- Change signature of `Session::call` to take `points_limit`
+- Change signature of `Session::call_raw` to take `points_limit`
+- Change signature of `Session::call` to return `CallReceipt`
+- Change signature of `Session::call_raw` to return `CallReceipt`
+
+### Removed
+
+- Remove `Session::set_point_limit`
+- Remove `Session::take_events`
+- Remove `Session::spent`
+
 ## [0.6.2] - 2023-07-07
 
 ### Added